+++ docs/CHANGELOG.md
@@ -6,10 +6,11 @@ by Matt Tropiano et al. (see AUTHORS.txt)
 Changed in [NOW]
 -----------------
 
-- `Added` BinaryObject file-reading methods for convenience.
-- `Added` TextObject file-reading methods for convenience.
 - `Fixed` GraphicUtils did not set transparent pixels when converted from full color images.
 - `Fixed` Some GraphicUtils method documentation was either misleading or unclear.
+- `Fixed` Tall Patch Graphics are now supported. (Issue #9)
+- `Added` BinaryObject file-reading methods for convenience.
+- `Added` TextObject file-reading methods for convenience.
 
 
 Changed in 2.12.0